refactor: refactorizacion de codigo repetido con la funcion handleAlternativesSubmit
